+++ src/main/java/com/sandbox/domain/todos/dto/TodoReq.java
@@ -1,0 +1,14 @@
+package com.sandbox.domain.todos.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class TodoReq {
+    private String content;
+}
